Remove oh my zsh.

+++ link.rb
@@ -1,18 +1,32 @@
 #!/usr/bin/env ruby
 
-require 'pathname'
-require 'yaml'
+require "pathname"
 
 home = Pathname.new(File.expand_path("~"))
 src_root = Pathname.new(File.dirname(__FILE__))
 
-link_spec_path = src_root + 'link.yaml'
-YAML.load_file(link_spec_path.realpath).each_pair do |target, source|
+files = {
+  ".tmux.conf" => "tmux.conf",
+  ".emacs.d" => "emacs.d",
+  ".spacemacs" => "spacemacs",
+  ".spacemacs-layers" => "spacemacs-layers",
+  ".gitconfig" => "gitconfig",
+  ".gitignore" => "gitignore",
+  ".zshrc" => "zshrc",
+  ".zpath" => "zpath",
+  ".zprofile" => "zprofile",
+  ".zshenv" => "zshenv",
+  ".config/bin" => "bin",
+  ".lein" => "lein",
+  ".gdbinit" => "gdbinit",
+  ".irbrc" => "irbrc",
+  ".rubocop.yml" => "rubocop.yml",
+}
+
+files.each do |target, source|
   target_path = home + target
 
-  if target_path.symlink?
-    target_path.unlink
-  end
+  target_path.unlink if target_path.symlink?
 
   if target_path.exist?
     puts "#{target_path} already exists."
